+++ R/comparisons.R
@@ -31,16 +31,18 @@
 #'     * "minmax": The highest and lowest levels of a factor.
 #'     * "revpairwise", "revreference", "revsequential": inverse of the corresponding hypotheses.
 #'     * Vector of length 2 with the two values to compare.
-#'     * Data frame with the same number of rows as `newdata`, with two columns of "low" and "high" values to compare.
-#'     * Function which accepts a numeric vector and returns a data frame with two columns of "low" and "high" values to compare. See examples below.
+#'     * Data frame with the same number of rows as `newdata`, with two columns of "lo" and "hi" values to compare.
+#'     * Function that accepts a vector and returns a data frame with two columns of "lo" and "hi" values to compare. See examples below.
 #'   - Logical variables:
 #'     * NULL: contrast between TRUE and FALSE
+#'     * Data frame with the same number of rows as `newdata`, with two columns of "lo" and "hi" values to compare.
+#'     * Function that accepts a vector and returns a data frame with two columns of "lo" and "hi" values to compare. See examples below.
 #'   - Numeric variables:
 #'     * Numeric of length 1: Contrast for a gap of `x`, computed at the observed value plus and minus `x / 2`. For example, estimating a `+1` contrast compares adjusted predictions when the regressor is equal to its observed value minus 0.5 and its observed value plus 0.5.
 #'     * Numeric of length equal to the number of rows in `newdata`: Same as above, but the contrast can be customized for each row of `newdata`.
 #'     * Numeric vector of length 2: Contrast between the 2nd element and the 1st element of the `x` vector.
-#'     * Data frame with the same number of rows as `newdata`, with two columns of "low" and "high" values to compare.
-#'     * Function which accepts a numeric vector and returns a data frame with two columns of "low" and "high" values to compare. See examples below.
+#'     * Data frame with the same number of rows as `newdata`, with two columns of "lo" and "hi" values to compare.
+#'     * Function that accepts a vector and returns a data frame with two columns of "lo" and "hi" values to compare. See examples below.
 #'     * "iqr": Contrast across the interquartile range of the regressor.
 #'     * "sd": Contrast across one standard deviation around the regressor mean.
 #'     * "2sd": Contrast across two standard deviations around the regressor mean.