+++ R/get_contrast_data_logical.R
@@ -2,6 +2,13 @@ get_contrast_data_logical <- function(model,
                                       newdata,
                                       variable,
                                       ...) {
+
+    # custom data frame or function
+    if (isTRUE(checkmate::check_function(variable$value)) || isTRUE(checkmate::check_data_frame(variable$value))) {
+        out <- contrast_categories_custom(variable, newdata)
+        return(out)
+    }
+
     lo <- hi <- newdata
     lo[[variable$name]] <- FALSE
     hi[[variable$name]] <- TRUE